Adding --method flag
Heavily refactored the vulnerable_versions_lookup function,
this improved readability, but also allowed to introduce the --method
flag, with this method we can specify which method(s) to use, or even use
all methods for search.

Now everything that the old decret version can do, the new one can as
well.